send notification mail if vserver tarball has wrong date

+++ centralise-backups.sh
@@ -9,6 +9,7 @@ vservers=$(/usr/sbin/vserver-stat | /usr/bin/awk '$1!="0" && NR>1  { print $8 }'
 
 for vserver in $vservers; do
         backup_file=$(find /etc/vservers/${vserver}/vdir/data/backup -maxdepth 1 -type f)
+	if [[ $backup_file != *$(date +%Y%m%d)*.sql.gz ]]; then echo "wrong vserver backup file name: $backup_file in centralise backups"| mail -s "ERROR vserver backup $backup_file" $admin_email ; fi
         [[ -n $backup_file ]] &&  cp $backup_file $backup_base
 
         # save the mysql dump seperately, easier to catch errors
